# app/static/makeascene.openapi.yaml
openapi: 3.1.0
info:
  title: makeascene
  description: |
    The API used for the frontend and third-party for Make A Scene
  version: 1.3.0
...
paths:
...
  /auth/email/confirm:
    get:
      summary: Confirm Email
      deprecated: false
      description: Used to confirm an account for creation via a sent email
      operationId: confirmPassword
      tags:
        - Authentication
      parameters:
        - name: token
          in: query
          description: ''
          required: true
          example: _sahjdoDSljjwpioSj1df...
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailConfirmed'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/email/resend:
    post:
      summary: Resend Email
      deprecated: false
      description: Used to request a resend of the confirmation email
      operationId: resendEmail
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailResendRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSent'
              example:
                message: Email sent if an unconfirmed user with that email exists.
          headers: {}
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/password/reset:
    post:
      summary: Change Password
      deprecated: false
      description: Used when the password should be changed or was forgotten
      operationId: changePassword
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChanged'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/logout:
    post:
      summary: Logout
      deprecated: false
      description: Logs a user out. Meaning it invalidates the refresh tokens
      operationId: logout
      tags:
        - Authentication
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
              example: >
                <!doctype html>

                <html lang=en>

                <title>404 Not Found</title>

                <h1>Not Found</h1>

                <p>The requested URL was not found on the server. If you entered
                the URL manually please check your spelling and try again.</p>
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
  /auth/password/request-reset:
    post:
      summary: Request Password Reset
      deprecated: false
      description: Used to request a password change with an email.
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailResendRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSent'
          headers: {}
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/sessions:
    post:
      summary: Login email and password/refresh token
      deprecated: false
      description: >-
        Authenticates a user with email and password or refresh token and
        returns a JWT token on success.
      operationId: localLogin
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    email:
                      type: string
                      format: email
                      examples:
                        - max.mustermann@example.com
                    password:
                      type: string
                      format: password
                      examples:
                        - securePassword123
                  required:
                    - email
                    - password
                - type: object
                  properties:
                    refresh_token:
                      type: string
                      examples:
                        - _sahjdoDSljjwpioSj1dfhj...
                  required:
                    - refresh_token
            examples:
              ValidLogin:
                value:
                  email: max.mustermann@example.com
                  password: securePassword123
                summary: Valid credentials
              WrongPassword:
                value:
                  email: max.mustermann@example.com
                  password: wrongPassword
                summary: Wrong password (triggers 401)
              UnknownUser:
                value:
                  email: unbekannt@example.com
                  password: securePassword123
                summary: Unknown email (triggers 401)
        required: true
      responses:
        '201':
          description: Login successful – JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              example:
                message: Login successful. Use token for authentication.
                access_token: eyJhbGciOiJIUzI1NiIsI...
                refresh_token: _sahjdoDSljjwpioSj1dfhj...
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/oauth/exchange:
    post:
      summary: Exchange login info for oauth users
      deprecated: false
      description: 'Gets the refresh and access tokens for a user that logs in with oauth. '
      operationId: exchangeOauth
      tags:
        - OAuth
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeTokenResponse'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/oauth/redirect:
    get:
      summary: Redirect to OAuth provider
      deprecated: false
      description: >
        Starts the OAuth 2.0 flow. The client is redirected to the chosen
        provider's login page. Currently only `google` is supported. After
        successful authentication the provider calls back to
        `/auth/oauth/google/callback`.
      operationId: oauthRedirect
      tags:
        - OAuth
        - Authentication
      parameters:
        - name: auth
          in: query
          description: ''
          required: true
          example: google
          schema:
            type: string
            enum:
              - google
      responses:
        '302':
          description: Redirect to OAuth provider login page (e.g. Google)
          content:
            text/html:
              schema:
                type: object
                properties: {}
          headers:
            Location:
              required: false
              description: URL of the OAuth provider's authorization endpoint
              schema:
                type: string
                example: >-
                  https://accounts.google.com/o/oauth2/v2/auth?client_id=...&redirect_uri=...&response_type=code&scope=openid+email+profile&state=...
        '400':
          description: Invalid or missing `auth` query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid authentication provider
                message: 'Supported providers: google'
          headers: {}
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
  /auth/oauth/google/callback:
    get:
      summary: OAuth callback (Google)
      deprecated: false
      description: >
        Called by Google after the user has authenticated. Exchanges the
        authorization code for a JWT token. Do not call this endpoint directly.
      tags:
        - OAuth
        - Authentication
      parameters:
        - name: code
          in: query
          description: Authorization code returned by Google
          required: false
          schema:
            type: string
            examples:
              - 4/0AX4XfWh...
        - name: state
          in: query
          description: >-
            State parameter for CSRF protection (must match the value set during
            redirect)
          required: false
          schema:
            type: string
            examples:
              - randomStateValue123
        - name: error
          in: query
          description: Set by Google if the user denied access
          required: false
          schema:
            type: string
            examples:
              - access_denied
      responses:
        '201':
          description: OAuth login successful – JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeTokenResponse'
              example:
                exchange: eyJhbGciOiJIUzI1NiIsI...
          headers: {}
        '400':
          description: OAuth failure (state mismatch, missing token, or missing user info)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                StateMismatch:
                  summary: CSRF state mismatch
                  value:
                    error: oauth_state_mismatch
                    message: Login session expired or invalid state. Please try again.
                TokenFailure:
                  summary: Google did not return a token
                  value:
                    error: oauth_failure
                    message: Failed to authorize with Google.
                UserInfoMissing:
                  summary: User info missing from token
                  value:
                    error: oauth_failure
                    message: Failed to fetch user info from Google.
                AuthFailure:
                  summary: Internal user authentication failed
                  value:
                    error: oauth_failure
                    message: Failed to authenticate user.
          headers: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
...
components:
  schemas:
    EmailConfirmed:
      type: object
      properties:
        message:
          type: string
          examples:
            - Email confirmed.
      required:
        - message
    EmailSent:
      type: object
      properties:
        message:
          type: string
          examples:
            - Email sent if an unconfirmed user with that email exists.
      required:
        - message
    EmailResendRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          examples:
            - test@test.com
      required:
        - email
    ChangePasswordRequest:
      type: object
      properties:
        token:
          type: string
          examples:
            - ltmk1VomO8dTxdKQEVYC39TKtCzU...
        new_password:
          type: string
          minLength: 8
          examples:
            - securePass123
          format: password
      required:
        - token
        - new_password
    PasswordChanged:
      type: object
      properties:
        message:
          type: string
          examples:
            - Password was changed
      required:
        - message
...
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
          examples:
            - All refresh tokens were invalidated.
      required:
        - message
    ExchangeTokenResponse:
      type: object
      properties:
        exchange:
          type: string
          examples:
            - sod2_02Jdj...
          description: |
            The token returned to retrieve the actual login data
      required:
        - exchange
...
    SessionResponse:
      type: object
      properties:
        message:
          type: string
          examples:
            - Login successful. Use token for authentication.
        access_token:
          type: string
          examples:
            - eyJhbGciOiJIUzI1NiIsI...
        refresh_token:
          type: string
          examples:
            - _sahjdoDSljjwpioSj1dfhj...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          examples:
            - Request body validation failed
        message:
          type: string
          examples:
            - A human-readable description of the error.
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                examples:
                  - unauthorized
              message:
                type: string
                examples:
                  - Invalid credentials.
            required:
              - error
              - message
          examples:
            Success:
              summary: Success
              value:
                error: unauthorized
                message: Invalid credentials.
    UnprocessableEntity:
      description: Validation error – one or more fields failed validation
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                examples:
                  - Request body validation failed
              fields:
                type: array
                items:
                  type: string
                examples:
                  - - '''password'' is too short (minimum length: 8)'
                    - '''email'' is not a valid email address'
          examples:
            Example 1:
              summary: Example 1
              value:
                error: Request body validation failed
                fields:
                  - '''password'' is too short (minimum length: 8)'
                  - '''email'' is not a valid email address'
    InternalServerError:
      description: ''
      content:
        application/octet-stream:
          schema:
            type: object
            properties: {}
          examples:
            Example 1:
              summary: Example 1
              value: {}
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
  - url: http://127.0.0.1:5000/v1
    description: http://127.0.0.1:5000/v1
security:
  - BearerAuth: []